feat(core): BEE-22 refactor to instance-based API eliminating all global state
- Create BeepingConfig value struct + compute_config() pure function
- Replace CBeepingCore with BeepingContext owning per-instance config
- Inject const BeepingConfig& into all Encoder/Decoder constructors
- Remove Globals::init() and all 28 mutable global variables
- Parameterize getToneFromIdx*/getFreqsFromIdx*/getLoudnessFromIdx
- Promote token/tone counts to constexpr in Globals namespace
- Make Encoder/Decoder destructors virtual
- Fix all -Wall -Wextra -Wpedantic warnings (0 remaining)
- Expand ThreadSafetyStressTest to cover all 7 modes
- Add ConcurrentInstancesTest (8 threads × 4 instances cross-mode)
- TSan CI now hard gate (removed continue-on-error)

Public C API (BeepingCoreLib_api.h) unchanged — full backwards compat.

+++ src/ReedSolomon.cpp
@@ -400,7 +400,7 @@ void ReedSolomon::GetCode(std::vector<int>& code) {
 
 // decoding received code
 void ReedSolomon::SetCode(const std::vector<int> code) {
-  if ((code.size() < kk) && (msg_len < kk)) {
+  if ((static_cast<int>(code.size()) < kk) && (msg_len < kk)) {
     // copy received error code digits (8)
     for (int i = 0; i < (nn - kk); ++i) recd[i] = code[msg_len + i];
